Rough tiles to block out TetraForce maps
I drew some 8x8 sprites as a tile map in Piskel and remade the map
in Tiled as proposed by TheRetroDragon. Then I implemented my
suggested changes and additions.
The layers are a bit messed up, especially the screen layer that I
could not get to work the way I wanted it to look like.

https://www.piskelapp.com/